+++ trunk/src/sreport/sreport.h
@@ -102,7 +102,7 @@ extern void *db_conn;
 extern uint32_t my_uid;
 
 extern void sreport_print_time(type_t type, print_field_t *field,
-			       uint32_t value, uint32_t total_time);
+			       uint64_t value, uint64_t total_time);
 extern int parse_option_end(char *option);
 extern char *strip_quotes(char *option, int *increased);
 extern void addto_char_list(List char_list, char *names);
